# Improve performance of empty_rows()