--- a/R/remove_empty.R
+++ b/R/remove_empty.R
@@ -77,7 +77,7 @@ empty_rows <- function(x) {
   if ((!is.matrix(x) && !is.data.frame(x)) || nrow(x) < 2) {
     vector("numeric")
   } else {
-    which(rowSums(is.na(x) | apply(x, c(1, 2), function(i) nchar(as.character(i)) == 0)) == ncol(x))
+    which(apply(x, 1, function(i) all(is.na(i) | nchar(as.character(i)) == 0)))
   }
 }